Refactor the tcx-constants
