+++ tcx-constants/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tcx-constants"
+version = "0.1.0"
+authors = ["Neal Xu <imxuneal@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+failure = "0.1.5"
+lazy_static = "1.3.0"
+serde_json = "1.0"
+serde = { version = "1.0", features = ["derive"] }
