+++ tcx-primitive/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+tcx-constants = { path = "../tcx-constants"}
 bitcoin = "0.20.0"
 byteorder = "1.2"
 secp256k1 = {version ="0.15", features = ["recovery"] }
