+++ tcx-tron/Cargo.toml
@@ -8,6 +8,7 @@ edition = "2018"
 tcx-crypto = { path = "../tcx-crypto" }
 tcx-chain = { path = "../tcx-chain" }
 tcx-primitive = { path = "../tcx-primitive" }
+tcx-constants = { path = "../tcx-constants" }
 
 hex-literal = "0.1.4"
 keccak-hash = "0.2.0"
