+++ tcx/src/lib.rs
@@ -24,9 +24,9 @@ use tcx_btc_fork::{
 };
 use tcx_chain::keystore::EmptyExtra;
 use tcx_chain::signer::TransactionSigner;
-use tcx_chain::{CoinInfo, HdKeystore, Metadata, TxSignResult};
+use tcx_chain::{HdKeystore, Metadata, TxSignResult};
+use tcx_constants::coin_info::{coin_info_from_symbol, coin_symbol_with_network};
 use tcx_crypto::{XPUB_COMMON_IV, XPUB_COMMON_KEY_128};
-use tcx_primitive::CurveType;
 use tcx_tron::{TrxAddress, TrxSignedTransaction, TrxTransaction};
 
 use std::convert::TryFrom;
@@ -85,75 +85,6 @@ fn delete_keystore_file(wid: &str) -> Result<()> {
     let path = Path::new(&ks_path);
     fs::remove_file(path)?;
     Ok(())
-}
-
-fn coin_info_from_symbol(symbol: &str) -> Result<CoinInfo> {
-    match symbol.to_uppercase().as_str() {
-        "BITCOINCASH" => Ok(CoinInfo {
-            symbol: "BITCOINCASH".to_string(),
-            derivation_path: "m/44'/145'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        "BITCOINCASH-TESTNET" => Ok(CoinInfo {
-            symbol: "BITCOINCASH".to_string(),
-            derivation_path: "m/44'/1'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        "LITECOIN" => Ok(CoinInfo {
-            symbol: "LITECOIN".to_string(),
-            derivation_path: "m/44'/2'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        "LITECOIN-P2WPKH" => Ok(CoinInfo {
-            symbol: "LITECOIN-P2WPKH".to_string(),
-            derivation_path: "m/44'/2'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        "LITECOIN-TESTNET" => Ok(CoinInfo {
-            symbol: "LITECOIN-TESTNET".to_string(),
-            derivation_path: "m/44'/1'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        "LITECOIN-TESTNET-P2WPKH" => Ok(CoinInfo {
-            symbol: "LITECOIN-TESTNET-P2WPKH".to_string(),
-            derivation_path: "m/44'/1'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        "TRON" => Ok(CoinInfo {
-            symbol: "TRON".to_string(),
-            derivation_path: "m/44'/195'/0'/0/0".to_string(),
-            curve: CurveType::SECP256k1,
-        }),
-        _ => Err(format_err!("unsupported_chain")),
-    }
-}
-
-const NETWORK_COINS: [&str; 3] = ["BITCOINCASH", "LITECOIN", "BITCOIN"];
-
-fn coin_symbol_with_network(v: &Value) -> String {
-    let chain_type = v["chainType"].as_str().expect("chainType");
-    if !NETWORK_COINS.contains(&chain_type) {
-        return chain_type.to_string();
-    }
-    let mut symbol = chain_type.to_string();
-
-    if let Some(network) = v["network"].as_str() {
-        if network.to_uppercase() != "MAINNET" {
-            symbol = format!("{}-{}", symbol, network);
-        }
-    }
-    if let Some(chain_id) = v["chainId"].as_str() {
-        if chain_id == "1" {
-            symbol = format!("{}-TESTNET", symbol);
-        }
-    }
-
-    if let Some(seg_wit) = v["segWit"].as_str() {
-        if seg_wit.to_uppercase() != "NONE" {
-            symbol = format!("{}-{}", symbol, seg_wit);
-        }
-    }
-    symbol
 }
 
 #[no_mangle]
